SSPM: Moved benchmarks to new APP make setup.
