+++ c/apps/sspm/.gitignore
@@ -1,0 +1,1 @@
+sspm.elf
