+++ c/apps/sspm/atomic.c
