+++ c/apps/sspm/atomic.h
