+++ c/apps/sspm/noc_roundtrip_bench.c
@@ -3,29 +3,27 @@
 #include <stdlib.h>
 #include <machine/patmos.h>
 #include <machine/rtc.h>
-#include "libcorethread/corethread.c"
-#include "libmp/mp.h"
-#include "libmp/mp_internal.h"
-#include "libsspm/sspm_properties.h"
+#include "../../libcorethread/corethread.h"
+#include "../../libmp/mp.h"
+#include "../../libmp/mp_internal.h"
 
 #define MP_CHAN_NUM_BUF 2
 #define MP_CHAN_BUF_SIZE 40
+#define CHANNEL_BUFFER_CAPACITY (256)
 
 const int NOC_MASTER = 0;
 
-const int CHANNEL_BUFFER_CAPACITY = 64;
 const int TIMES_TO_SEND = 1000;
 volatile _UNCACHED int send_clock[TIMES_TO_SEND];
 volatile _UNCACHED int recv_clock[TIMES_TO_SEND];
 volatile _UNCACHED int ack_recv_clock[TIMES_TO_SEND];
 
-typedef enum{
-	TRANSMIT,
-	ACKNOWLEDGE
-} TRANSMISSION_STATE;
 
 void sender_slave(void* args){
-	volatile _SPM int* flag = (volatile _SPM int*) LOWEST_SSPM_ADDRESS;
+	int cpuid = get_cpuid();
+	qpd_t * chan = mp_create_qport(1, SOURCE, CHANNEL_BUFFER_CAPACITY*sizeof(int),MP_CHAN_NUM_BUF);
+
+	mp_init_ports();
 
 	int send, ack_recv;
 
@@ -34,8 +32,9 @@ void sender_slave(void* args){
 		send = get_cpu_cycles();
 		asm volatile ("" : : : "memory");
 
-		*flag = TRANSMIT;
-		while(*flag != ACKNOWLEDGE){}
+		mp_send(chan,0);
+		
+		while(*(chan->send_recv_count) != (k+1)){}
 		
 		asm volatile ("" : : : "memory");
 		ack_recv = get_cpu_cycles();
@@ -48,17 +47,20 @@ void sender_slave(void* args){
 
 void receiver_slave(void* args){
 	int cpuid = get_cpuid();
-	volatile _SPM int* flag = (volatile _SPM int*) LOWEST_SSPM_ADDRESS;
+	qpd_t * chan = mp_create_qport(1, SINK, CHANNEL_BUFFER_CAPACITY*sizeof(int), MP_CHAN_NUM_BUF);
+	
+	mp_init_ports();
 
 	int recv;
 	
 	for(int k = 0; k<TIMES_TO_SEND; k++){
-		while(*flag != TRANSMIT){}
+		mp_recv(chan,0);
+		
 		asm volatile ("" : : : "memory");
 		recv = get_cpu_cycles();
 		asm volatile ("" : : : "memory");
 
-		*flag = ACKNOWLEDGE;
+		mp_ack(chan,0);
 
 		asm volatile ("" : : : "memory");
 		recv_clock[k] = recv;
@@ -68,13 +70,11 @@ void receiver_slave(void* args){
 
 
 int main(){
-	volatile _SPM int* flag = (volatile _SPM int*) LOWEST_SSPM_ADDRESS;
-	*flag = ACKNOWLEDGE;
 
 	int core1 = 1;
 	int core2 = 2;
-	corethread_create(&core1, &sender_slave, NULL);
-	corethread_create(&core2, &receiver_slave, NULL);
+	corethread_create(core1, &sender_slave, NULL);
+	corethread_create(core2, &receiver_slave, NULL);
 
 	int res;
 	corethread_join(core1, (void **) &res);
