+++ c/apps/sspm/sspm_properties.h
