--- c/libsspm/benchmarks/README.md
@@ -1,7 +1,0 @@
-
-This folder contains benchmarking programs for the SSPM.
-
-Each file is an independant benchmark. To run it, move/copy the file into the "patmos/c" folder
-and run it like any other program.
-
-
